+++ examples/SimpleExample.hs
@@ -5,14 +5,16 @@
 {-# LANGUAGE GADTs             #-}
 module Main where
 
-import qualified Knit.Report as K    
+import qualified Knit.Report              as K    
 
-import qualified Data.Map                      as M
-import qualified Data.Text.IO                  as T
-import qualified Data.Text.Lazy                as TL
+import qualified Data.Map                 as M
+import qualified Data.Text.IO             as T
+import qualified Data.Text.Lazy           as TL
 import qualified Data.Text                as T
-import           Data.String.Here (here)
-import qualified Graphics.Vega.VegaLite        as V
+import           Data.String.Here          (here)
+import qualified Graphics.Vega.VegaLite   as V
+
+import qualified Plots                    as P
 
 templateVars :: M.Map String String
 templateVars = M.fromList
@@ -42,10 +44,7 @@ md1 = [here|
 [MarkDownLink]:<https://pandoc.org/MANUAL.html#pandocs-markdown>
 |]
 
-makeDoc :: (K.Member K.ToPandoc effs -- required for the single-document variant
-           , K.PandocEffects effs -- all effects for knitting
-           ) 
-        => K.Sem effs ()
+makeDoc :: K.KnitOne effs => K.Sem effs ()
 makeDoc = K.wrapPrefix "makeDoc" $ do
   K.logLE K.Info "adding some markdown..."
   K.addMarkDown md1
@@ -54,8 +53,13 @@ makeDoc = K.wrapPrefix "makeDoc" $ do
   K.addLatex "Overused favorite equation: $e^{i\\pi} + 1 = 0$"
   K.logLE K.Info "adding a visualization..."
   K.addMarkDown "## An example hvega visualization"
-  K.addHvega "someID" exampleVis
+  K.addHvega Nothing (Just "From the cars data-set") exampleVis
+  K.addMarkDown "## An example Diagrams visualization"
+  K.logLE K.Info "adding a Diagrams plot..."  
+  K.addDiagramAsSVG Nothing (Just "Example diagrams visualization using the Plots library") 300 300 samplePlot
 
+
+-- example using HVega  
 exampleVis :: V.VegaLite
 exampleVis =
   let cars =  V.dataFromUrl "https://vega.github.io/vega-datasets/data/cars.json" []
@@ -65,3 +69,22 @@ exampleVis =
         . V.color [ V.MName "Origin", V.MmType V.Nominal ]
       bkg = V.background "rgba(0, 0, 0, 0.05)"
   in V.toVegaLite [ bkg, cars, V.mark V.Circle [], enc [] ]  
+
+
+-- example using Plots (as an example of using Diagrams)
+samplePlot :: K.Diagram K.SVG
+samplePlot = P.renderAxis logAxis
+
+logData = [K.V2 1 10, K.V2 2 100, K.V2 2.5 316, K.V2 3 1000]
+
+logAxis :: P.Axis K.SVG K.V2 Double
+logAxis = P.r2Axis K.&~ do
+  P.scatterPlot' logData
+  -- yMin ?= 200
+
+  P.yAxis P.&= do
+    P.logScale K..= P.LogAxis
+    P.majorTicksFunction K..= P.logMajorTicks 5 -- <> pure [1]
+    -- minorTicksFunction .= minorTicksHelper 5
+
+--exampleDiagram ::